Update medical AI consultation system with enhanced features

+++ backend/models/patient.py
@@ -15,7 +15,7 @@ class Symptoms(BaseModel):
 
 class BiologicalAnalysis(BaseModel):
     crp: float
-    leukocytes: int
+    leukocytes: float
     creatinine: float
     asat: float
     alat: float
